refactor(generation,bin-designer): split bridge types + lift lid-section helpers
Behavior-preserving. bridge/types.ts becomes a barrel re-exporting
messages.ts (requests), responses.ts (responses), and meshData.ts (mesh
shapes); every export name is preserved so importers are unaffected. The
pure lid-section helpers move to useLidSection.helpers.ts (with a test),
re-exported from useLidSection.ts for existing importers.
